Garbage Collection bug fixed for communicator TMap objects

+++ Source/StoneAgeColony/Communicator.h
@@ -51,7 +51,7 @@ public:
 	int test;
 
 	// IDs of usable items in game
-	TMap<unsigned short, AUsableActor*> UsableItemIDMap; // how?
+	TMap<int, AUsableActor*> UsableItemIDMap; // how?
 	
 	// TArray of ints holding ids of items in player inventory
 	TArray<int> PlayerInventory;
